Add factories

+++ src/main/java/org/chainoptim/desktop/shared/caching/CachingService.java
@@ -2,7 +2,7 @@ package org.chainoptim.desktop.shared.caching;
 
 public interface CachingService<T> {
 
-    void add(String key, T data, long staleTime);
+    void add(String key, T data, float staleTime);
     T get(String key);
     boolean isCached(String key);
     boolean isStale(String key);
